requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cx-Oracle==8.3.0
et-xmlfile==1.1.0
fastapi==0.79.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
pydantic==1.9.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
requests==2.28.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.2
xlrd==2.0.1
XlsxWriter==3.0.3
